requirements.txt
asgiref==3.8.1
attrs==23.2.0
blinker==1.9.0
branca==0.8.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.1
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.12.1
dbfread==2.0.7
Django==4.2
django-colorfield==0.11.0
django-colorful==1.3
django-jazzmin==3.0.1
django-matplotlib==0.1
django-model-utils==5.0.0
django-multiupload==0.6.1
django-notifications-hq==1.8.3
django-simple-history==3.7.0
django-unfold==0.43.0
fastkml==0.12
fiona==1.9.6
Flask==3.1.0
folium==0.18.0
fonttools==4.55.3
geopandas==1.0.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonfield==3.1.0
kiwisolver==1.4.8
kmz-custom-maps==0.1.0
lxml==4.9.4
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
pygeoif==0.7
pykml==0.2.0
pyogrio==0.10.0
pyparsing==3.2.1
pyphen==0.17.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
shapely==2.0.6
simplekml==1.3.6
six==1.16.0
sqlparse==0.5.0
swapper==1.4.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
weasyprint==63.1
webencodings==0.5.1
Werkzeug==3.1.3
xyzservices==2024.4.0
zipfile36==0.1.3
zopfli==0.2.3.post1
